chore: fix pnpm warnings when package-lock = false
I kept getting the "WARN  A pnpm-lock.yaml file exists. The current configuration prohibits to read or write a lockfile" warning with pnpm and it was because I had this set to false at the profile level (I work mostly on modules, and keep it off by default). As long as we are using lock files in this repo, let's keep this setting pinned down correctly.